Be silent when injecting into process
In the release version, we remove some logs (except error level) added into the injected process.
This helps to leave fewer unnecessary logs which are impossible for modules to hide.

This function can be toggled using the static field `beSilent`.
Developers are encouraged to use the debug version of LSPosed to develop modules.